refactor: Move file utils into utils directory
